Changed current from a float to a uint8_t.  This may end up dropping partial amps but i've never actually seen partial amps in practice

+++ io.h
@@ -27,7 +27,7 @@ class TeslaControllerIO {
         virtual void onDebugMessage(std::function<void(bool)>) = 0;
         virtual void writeRaw(uint8_t*, size_t) = 0;
         virtual void writeRawPacket(uint8_t *data, size_t length) = 0;
-        virtual void writeActualCurrent(float) = 0;
+        virtual void writeActualCurrent(uint8_t) = 0;
         virtual void writeCharger(uint16_t, uint8_t) = 0;
         virtual void writeChargerTotalKwh(uint16_t, uint32_t) = 0;
         virtual void writeChargerSerial(uint16_t, uint8_t*, size_t length) = 0;
