+++ twc_connector.h
@@ -24,8 +24,8 @@ class TeslaConnector {
         TeslaConnector(uint16_t twcid, uint8_t max_charge_rate);
         void SetVin(uint8_t* upperVin);
         uint8_t* GetVin();
-        void SetActualCurrent(float current);
-        float GetActualCurrent();
+        void SetActualCurrent(uint8_t current);
+        uint8_t GetActualCurrent();
         uint8_t GetPhaseCurrent(uint8_t phase);
         
         uint16_t twcid;
@@ -46,7 +46,7 @@ class TeslaConnector {
         uint8_t max_allowable_current;
     private:
         uint8_t vin_[18];
-        float actual_current_;
+        uint8_t actual_current_;
 };
 
 #endif /* TWC_CONNECTOR_H */
